Code re-organisation
Stubs are now in their own seperate file, and are reorganized so they
are easier to automate when the time comes. Each hardware now has
their own stub class which is a child of stubs.py:Stub, and the
mapping is defined in stubs.py:stub_dict.

Context now takes in the overlay, this simplifies the interfacing a
little bit. Register function is now also in Context rather than being
a seperate function.