WIP: Completed new stub implementation.
warn: This is a very intermediate commit

newstubs will soon replace stubs.py, and as of this commit is what's
used in pushpush.py

currently rigged to work just with the add module, and the json file
reflects what the rest of the files will look like as well

a bunch of TODOs, as well as refactoring todo. currently this is just
compatability code.